# 🍽 Combine integration test executables for faster builds