--- a/cli/tests/integration/body.rs
+++ b/cli/tests/integration/body.rs
@@ -1,9 +1,7 @@
 //! Tests related to HTTP request and response bodies.
 
-mod common;
-
 use {
-    common::{Test, TestResult},
+    crate::common::{Test, TestResult},
     hyper::{body, StatusCode},
 };