--- a/cli/tests/integration/downstream_req.rs
+++ b/cli/tests/integration/downstream_req.rs
@@ -1,7 +1,5 @@
-mod common;
-
 use {
-    common::{Test, TestResult},
+    crate::common::{Test, TestResult},
     hyper::{Request, StatusCode},
 };