--- a/cli/tests/integration/main.rs
+++ b/cli/tests/integration/main.rs
@@ -1,0 +1,13 @@
+mod body;
+mod common;
+mod downstream_req;
+mod env_vars;
+mod http_semantics;
+mod logging;
+mod request;
+mod response;
+mod sending_response;
+mod sleep;
+mod upstream;
+mod upstream_async;
+mod upstream_streaming;