--- a/cli/tests/integration/request.rs
+++ b/cli/tests/integration/request.rs
@@ -1,7 +1,5 @@
-mod common;
-
 use {
-    common::{Test, TestResult},
+    crate::common::{Test, TestResult},
     hyper::StatusCode,
 };