--- a/cli/tests/integration/response.rs
+++ b/cli/tests/integration/response.rs
@@ -1,7 +1,5 @@
-mod common;
-
 use {
-    common::{Test, TestResult},
+    crate::common::{Test, TestResult},
     hyper::StatusCode,
 };