--- a/cli/tests/integration/upstream_async.rs
+++ b/cli/tests/integration/upstream_async.rs
@@ -1,7 +1,5 @@
-mod common;
-
 use {
-    common::{Test, TestResult},
+    crate::common::{Test, TestResult},
     hyper::{Response, StatusCode},
 };