--- a/cli/tests/integration/upstream_streaming.rs
+++ b/cli/tests/integration/upstream_streaming.rs
@@ -1,7 +1,5 @@
-mod common;
-
 use {
-    common::{Test, TestResult},
+    crate::common::{Test, TestResult},
     hyper::{body::HttpBody, Response, StatusCode},
 };