… cli/tests/integration/common.rs