@acfoltzer 2022-02-16
This seems to sort of improve CI runtimes, but not as much as I would have expected. The `test` step in Ubuntu went from taking ~26 minutes on typical CI runs to ~18 minutes on this branch, but the Windows and Mac runs seem to have taken the same time. I'm really unsure how to account for this, but since it's not worse, I'm inclined to go ahead with this change and hope to understand those platforms better later.